requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord==1.0.1
discord-py-slash-command==1.0.9.4
discord.py==1.6.0
idna==3.1
multidict==5.1.0
mysql-connector-python==8.0.23
mysqlclient==2.0.3
protobuf==3.15.1
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
